ISSUE-345: move to dirs
