+++ tests/Integration/Domain/Repository/Fixtures/Identity/AdministratorTokenWithAdministrator.csv
