+++ tests/Integration/Domain/Repository/Fixtures/Messaging/MessageFixture.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PhpList\Core\Tests\Integration\Domain\Repository\Fixtures;
+namespace PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Messaging;
 
 use DateTime;
 use Doctrine\Bundle\FixturesBundle\Fixture;
