+++ tests/Integration/Domain/Repository/Identity/AdministratorRepositoryTest.php
@@ -10,7 +10,7 @@ use PhpList\Core\Domain\Model\Identity\Administrator;
 use PhpList\Core\Domain\Repository\Identity\AdministratorRepository;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
 use PhpList\Core\TestingSupport\Traits\ModelTestTrait;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Identity\AdministratorFixture;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
 /**
