+++ tests/Integration/Domain/Repository/Messaging/SubscriberListRepositoryTest.php
@@ -15,10 +15,10 @@ use PhpList\Core\Domain\Repository\Subscription\SubscriberRepository;
 use PhpList\Core\Domain\Repository\Subscription\SubscriptionRepository;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
 use PhpList\Core\TestingSupport\Traits\SimilarDatesAssertionTrait;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorFixture;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\SubscriberFixture;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\SubscriberListFixture;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\SubscriptionFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Identity\AdministratorFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Subscription\SubscriberFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Subscription\SubscriberListFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Subscription\SubscriptionFixture;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 
 /**
