+++ tests/Integration/Security/AuthenticationTest.php
@@ -8,8 +8,8 @@ use Doctrine\ORM\Tools\SchemaTool;
 use PhpList\Core\Domain\Model\Identity\Administrator;
 use PhpList\Core\Security\Authentication;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorFixture;
-use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\AdministratorTokenWithAdministratorFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Identity\AdministratorFixture;
+use PhpList\Core\Tests\Integration\Domain\Repository\Fixtures\Identity\AdministratorTokenWithAdministratorFixture;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
 use Symfony\Component\HttpFoundation\Request;
